refactor: 提取缓冲处理逻辑到astr_message_event
